requirements.txt
Django==5.0.0
djangorestframework==3.14.0
psycopg2-binary==2.9.9
python-decouple==3.8
requests==2.31.0
django-cors-headers==4.3.1
supabase==2.3.4